Use flex grid view while listing collections
